% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_cssproperty_marginInput.R
\name{mod_cssproperty_marginInput}
\alias{mod_cssproperty_marginInput}
\alias{mod_cssproperty_margin}
\title{mod_cssproperty_marginInput and mod_cssproperty_margin}
\usage{
mod_cssproperty_marginInput(id, title)

mod_cssproperty_margin(input, output, session)
}
\arguments{
\item{id}{shiny id}

\item{title}{title of the \code{gradientBox}}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny module that allows to give css property for \code{@Page} rule in CSS for pages media
}
\examples{
\dontrun{
library(shiny)
library(shinydashboardPlus)
if (interactive()) {
ui <- fluidPage(
mod_cssproperty_marginInput("left-corner")
)

server <- function(input, output, session) {
callModule(mod_cssproperty_margin, "left-corner")
}

shinyApp(ui, server)
}
}
}
