% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_view_renderedpdfUI.R
\name{mod_view_renderedpdfUI}
\alias{mod_view_renderedpdfUI}
\alias{mod_view_renderedpdf}
\title{mod_view_renderedpdfUI and mod_view_renderedpdf}
\usage{
mod_view_renderedpdfUI(id)

mod_view_renderedpdf(input, output, session, path,
  filename = "template_report", paramsUI = reactive(NULL),
  style_css = reactive(NULL))
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{inernal}

\item{path}{character. Path to the file to render}

\item{filename}{string value. Name without extension of the file to render}

\item{paramsUI}{reactive. A list of named parameters that override custom params
specified within the YAML front-matter (see argument \code{params} in \code{render})}

\item{style_css}{css file}
}
\description{
A shiny module that displays a PDF file rendered from a Rmd file
}
\examples{
library(shiny)
if (interactive()) {
ui <- fluidPage(
mod_view_renderedpdfUI("my_pdf")
)

server <- function(input, output, session) {
callModule(mod_view_renderedpdf, "my_pdf", path = system.file("www", package = "wimpress"),
filename = "template_report")
}

shinyApp(ui, server)
}

}
