% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_cssproperty_pageInput.R
\name{mod_cssproperty_pageInput}
\alias{mod_cssproperty_pageInput}
\alias{mod_cssproperty_page}
\title{mod_cssproperty_pageInput and mod_cssproperty_page}
\usage{
mod_cssproperty_pageInput(id, title, col = "black")

mod_cssproperty_page(input, output, session)
}
\arguments{
\item{id}{shiny id}

\item{title}{title of the \code{gradientBox}}

\item{col}{background color of the \code{gradientBox}}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny module that allows to give css property for \code{@Page} rule in CSS for pages media
}
\examples{
\dontrun{
library(shiny)
library(shinydashboardPlus)
if (interactive()) {
ui <- fluidPage(
mod_cssproperty_pageInput("page", title = "page")
)

server <- function(input, output, session) {
callModule(mod_cssproperty_page, "page")
}

shinyApp(ui, server)
}
}
}
